Split `DataStorage.prototype.save` and made changes to loops.
Now `DataStorage.prototype.save(force)` is now divided into 2 individual methods, `DataStorage.prototype.save()` and `DataStorage.prototype.saveAsync()`.
Now loops will return a promise, and to break a loop a new instance of `BreakLoop` must be passed to it.
Renamed `ObjectNotation` to `NotationHandler` and instead of an anonymous function it is now a constructor.
Small changes to the rest of the code.